+++ app/components/Layout.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import React from 'react';
 import { Toaster } from 'react-hot-toast';
 
-import ToaasterLayout from './ToaasterLayout';
+import ToasterLayout from './ToasterLayout';
 
 export default function Layout({ children }) {
   return (
@@ -11,7 +11,7 @@ export default function Layout({ children }) {
         <meta charSet="utf-8" />
         <meta httpEquiv="Content-Type" content="text/html; charset=utf-8" />
       </Head>
-      <ToaasterLayout />
+      <ToasterLayout />
       <Toaster />
       {children}
     </>
